Simplify code and table
